fix: commonJs default export
