+++ index.cjs
@@ -1,1 +1,6 @@
-module.exports = require('./build/index.js')
+const sdk = require('./build/index.js')
+// patch default for builders expecting it downstream
+module.exports = {
+  ...sdk,
+  default: sdk.default ?? sdk
+};
