+++ integration-tests/node/cjs/package.json
@@ -2,6 +2,10 @@
   "name": "@identus-edge-agent-sdk/integration-tests-cjs",
   "version": "0.0.0",
   "private": true,
+  "type": "commonjs",
+  "scripts": {
+    "test": "node ./test.cjs"
+  },
   "devDependencies": {
     "sdk": "file:../../../"
   },
