+++ integration-tests/node/esm/package.json
@@ -3,6 +3,9 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
+  "scripts": {
+    "test": "node ./test.mjs"
+  },
   "devDependencies": {
     "sdk": "file:../../../"
   },
